cloud_functions/bq_to_txt/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.20.1
google-cloud-storage==2.17.0
pandas==2.2.2
